LayoutTests/accessibility/mac/AOM-relation-property.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../../resources/accessibility-helper.js"></script>
</head>
<body>

<div id="menu" role="menu" aria-label="First menu" tabindex="-1">
<div role="menuitem" id="item1" aria-label="item 1" tabindex="-1">Menu item 1</div>
<div role="menuitem" id="item2" aria-label="item 2" tabindex="0">Menu item 2</div>
</div>

<input id="input">
<div id="details">Detailed descriptive information.</div>
<div id="error">Information about the error.</div>

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests relation type Accessibility Object Model properties.");
    if (window.accessibilityController) {
        var node;
        var axNode;
        var relation;
        var axRelation;

        testActiveDescendant();
        testDetails();
        testErrorMessage();
    }

    function testActiveDescendant() {
        debug("\nActiveDescendant support");
        node = document.getElementById("menu");
        axNode = accessibilityController.accessibleElementById("menu");
        relation = document.getElementById("item1");
        axRelation = accessibilityController.accessibleElementById("item1");

        // Active Descendant makes the Menu item selected. 
        shouldBeFalse("axRelation.isSelected");
        node.accessibleNode.activeDescendant = relation.accessibleNode;
        shouldBeTrue("axRelation.isSelected");
        
        // Test that we are able to get the right container.
        shouldBeFalse("axRelation.boolAttributeValue('AXIsActiveDescendantOfFocusedContainer')");
        node.focus();
        shouldBeTrue("axRelation.boolAttributeValue('AXIsActiveDescendantOfFocusedContainer')");
        
        // Make sure it falls back to ARIA attribute if AOM property is not set.
        node.accessibleNode.activeDescendant = null;
        shouldBeFalse("axRelation.boolAttributeValue('AXIsActiveDescendantOfFocusedContainer')");
        node.setAttribute("aria-activedescendant", "item1");
        shouldBeTrue("axRelation.boolAttributeValue('AXIsActiveDescendantOfFocusedContainer')");
    }

    function testDetails() {
        debug("\nDetails support");
        node = document.getElementById("input");
        axNode = accessibilityController.accessibleElementById("input");
        relation = document.getElementById("details");
        
        shouldBeNull("axNode.ariaDetailsElementAtIndex(0)");
        node.accessibleNode.details = relation.accessibleNode;
        shouldBeEqualToString("axNode.ariaDetailsElementAtIndex(0).childAtIndex(0).stringValue", "AXValue: Detailed descriptive information.");
        
        // Test removing the details element.
        relation.parentElement.removeChild(relation);
        shouldBeNull("axNode.ariaDetailsElementAtIndex(0)");
    }

    function testErrorMessage() {
        debug("\nErrorMessage support");
        node = document.getElementById("input");
        axNode = accessibilityController.accessibleElementById("input");
        relation = document.getElementById("error");
        
        shouldBeNull("axNode.ariaErrorMessageElementAtIndex(0)");
        node.accessibleNode.errorMessage = relation.accessibleNode;
        shouldBeEqualToString("axNode.ariaErrorMessageElementAtIndex(0).childAtIndex(0).stringValue", "AXValue: Information about the error.");
    }

</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
