LayoutTests/accessibility/mac/AOM-relation-property-expected.txt

Menu item 1
Menu item 2

Information about the error.
This tests relation type Accessibility Object Model properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



ActiveDescendant support
PASS axRelation.isSelected is false
PASS axRelation.isSelected is true
PASS axRelation.boolAttributeValue('AXIsActiveDescendantOfFocusedContainer') is false
PASS axRelation.boolAttributeValue('AXIsActiveDescendantOfFocusedContainer') is true
PASS axRelation.boolAttributeValue('AXIsActiveDescendantOfFocusedContainer') is false
PASS axRelation.boolAttributeValue('AXIsActiveDescendantOfFocusedContainer') is true

Details support
PASS axNode.ariaDetailsElementAtIndex(0) is null
PASS axNode.ariaDetailsElementAtIndex(0).childAtIndex(0).stringValue is "AXValue: Detailed descriptive information."
PASS axNode.ariaDetailsElementAtIndex(0) is null

ErrorMessage support
PASS axNode.ariaErrorMessageElementAtIndex(0) is null
PASS axNode.ariaErrorMessageElementAtIndex(0).childAtIndex(0).stringValue is "AXValue: Information about the error."
